requirements.txt
pandas==2.3.3
numpy==2.4.0
ccxt==4.5.29
requests==2.32.5
yfinance==1.0
python-dotenv==1.2.1
